ブロックの一番高いポジション取得が間違っていたので直しました。

+++ mogupro/game/src/Game/Field/cBlock.cpp
@@ -10,9 +10,9 @@ namespace Field
 {
 cBlock::cBlock() :
     mPosition( vec3( 0 ) )
-    , mScale( BLOCK_SIZE )
+    , mScale( (float) BLOCK_SIZE )
     , mIsActive( false )
-    , mCollider( mPosition, vec3( BLOCK_SIZE ), vec3( 0.5f ) )
+    , mCollider( mPosition, vec3( (float) BLOCK_SIZE ), vec3( 0.5f ) )
 {
 
 }
